requirements.txt
streamlit==1.29.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.1
numpy==1.26.4